server/requirements.txt
signalfx
pyserial
sigfig